+++ changelog.md
@@ -1,3 +1,7 @@
-v0.8.0:
-- creating `~/.fency/runtime`, `~/.fency/libs` automatically
-- structs WIP, but at least works on stack
+v0.8.1:
+- added types `single`, `u32`, `i32`, `ubyte`, `ibyte`
+- seman: fields type mismatch
+- fixes
+- updated `examples/etc_raylib.fcy` example
+- conv ops for new types
+- address by struct field